Fixes the type of view port size.
This commit fixes the issue described in https://github.com/memononen/nanovg/pull/477.

+++ src/nanovg_mtl.m
@@ -1440,7 +1440,7 @@ static int mtlnvg__renderUpdateTexture(void* uptr, int image, int x, int y,
   return 1;
 }
 
-static void mtlnvg__renderViewport(void* uptr, int width, int height,
+static void mtlnvg__renderViewport(void* uptr, float width, float height,
                                    float devicePixelRatio) {
   MNVGcontext* mtl = (MNVGcontext*)uptr;
   mtl->viewPortSize = (vector_uint2){width * devicePixelRatio,
